requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
statsmodels==0.14.4
plotly==6.0.1
streamlit==1.42.0
pyyaml==6.0.2
openpyxl==3.1.5
pyarrow==19.0.1
polars==1.22.0
python-dateutil==2.9.0.post0
pytest==8.3.5
ruff==0.9.9
black==25.1.0
